Implemented Keyboard Store with product and coupon microservices.
Added Eureka Server for service registration and discovery.

Added Zuul proxy server to act as an API gateway for the microservices.

Added Coupon Microservice at add discounts to product prices.

Implemented discounts for products based on Coupon Microservice. Implemented using the FeignClient and Zuul API gateway to get the Coupon based on the couponCode in the product.

Added Hystrix to add fault tolerance to the Product microservice so that it can run even if the Coupon microservice is not available.

Added Hystrix Dashboard to visualize errors over time in product microservice. Configuration bean was used in product service to register hystrix.stream.

Added Sleuth to microservices to add traces across microservices requests.

Added Zipkin dependency to microservices to visualize trace requests on a dashboard between microservices for debugging.